authorization working n configured.

+++ routers/posts.py
@@ -10,6 +10,8 @@ import models
 from database import get_db
 from schemas import PostCreate, PostResponse, PostUpdate
 
+from auth import Current_User
+
 router = APIRouter()  # this is what will decorate our apps instead of the main app, allows us to organize our code better and separate concerns
 
 
@@ -30,21 +32,15 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
     response_model=PostResponse,
     status_code=status.HTTP_201_CREATED,
 )
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(
-        select(models.User).where(models.User.id == post.user_id),
-    )
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
-
+async def create_post(
+    post: PostCreate,
+    current_user: Current_User, 
+    db: Annotated[AsyncSession, Depends(get_db)]
+):
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id, #user id comes from authenticated user, not from the request body, ensures only logged in users can create posts and they can only create posts for themselves
     )
 
     db.add(new_post)
@@ -76,6 +72,7 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
 async def update_post_full(
     post_id: int,
     post_data: PostCreate,
+    current_user: Current_User,
     db: Annotated[AsyncSession, Depends(get_db)],
 ):
     result = await db.execute(
@@ -90,22 +87,15 @@ async def update_post_full(
             detail="Post not found",
         )
 
-    if post_data.user_id != post.user_id:
-        result = await db.execute(
-            select(models.User).where(models.User.id == post_data.user_id),
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post",
         )
-
-        user = result.scalars().first()
-
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found",
-            )
 
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
+    # post.user_id = post_data.user_id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
@@ -117,6 +107,7 @@ async def update_post_full(
 async def update_post_partial(
     post_id: int,
     post_data: PostUpdate,
+    current_user: Current_User,
     db: Annotated[AsyncSession, Depends(get_db)],
 ):
     result = await db.execute(
@@ -129,6 +120,12 @@ async def update_post_partial(
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not found",
+        )
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post",
         )
 
     update_data = post_data.model_dump(exclude_unset=True)
@@ -146,7 +143,11 @@ async def update_post_partial(
 
 
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(
+    post_id: int, 
+    current_user: Current_User, 
+    db: Annotated[AsyncSession, Depends(get_db)]
+):
     result = await db.execute(
         select(models.Post).where(models.Post.id == post_id),
     )
@@ -157,6 +158,11 @@ async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not found",
+        )
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to delete this post",
         )
 
     await db.delete(post)
